# BE/openapi.yaml
openapi: 3.0.3
info:
  title: Stock Management API
  description: API for managing inventory, stock-in, stock-out operations and reports
  version: 1.0.0
  contact:
    name: TCL Group
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
...
paths:
  /products:
    post:
      tags:
        - Products
      summary: Create a new product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
            example:
              sku: "SKU-001"
              name: "Product A"
              customer_id: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: "Success create data."
                data:
                  id: "123e4567-e89b-12d3-a456-426614174000"
                  sku: "SKU-001"
                  name: "Product A"
                  customer_id: "550e8400-e29b-41d4-a716-446655440000"
                  created_at: "2026-03-05T02:30:00Z"
                  updated_at: "2026-03-05T02:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags:
        - Products
      summary: List all products
      operationId: listProducts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
              example:
                success: true
                message: "Success get all data."
                data:
                  - id: "123e4567-e89b-12d3-a456-426614174000"
                    sku: "SKU-001"
                    name: "Product A"
                    customer_id: "550e8400-e29b-41d4-a716-446655440000"
                    created_at: "2026-03-05T02:30:00Z"
                    updated_at: "2026-03-05T02:30:00Z"
                pagination:
                  page: 1
                  per_page: 10
                  total: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (starts from 1)
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  schemas:
    CreateProductRequest:
      type: object
      required:
        - sku
        - name
        - customer_id
      properties:
        sku:
          type: string
          description: Product SKU
        name:
          type: string
          description: Product name
        customer_id:
          type: string
          format: uuid
          description: Customer UUID
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            type: object
        pagination:
          type: object
          properties:
            page:
              type: integer
            per_page:
              type: integer
            total:
              type: integer
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "INVALID_REQUEST"
              message: "Invalid request parameters"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "INTERNAL_ERROR"
              message: "An internal error occurred"